Consolidate utility classes into utils.py
Rather than having paging_wrapper and stub_response as separate
modules, just put them into utils.
