+++ src/globus_cli/commands/task/event_list.py
@@ -3,7 +3,6 @@ import json
 import click
 
 from globus_cli.login_manager import requires_login
-from globus_cli.paging_wrapper import PagingWrapper
 from globus_cli.parsing import command, task_id_arg
 from globus_cli.services.transfer import (
     TRANSFER_RESOURCE_SERVER,
@@ -11,6 +10,7 @@ from globus_cli.services.transfer import (
     iterable_response_to_dict,
 )
 from globus_cli.termio import formatted_print
+from globus_cli.utils import PagingWrapper
 
 
 @command(
