+++ src/globus_cli/utils.py
@@ -1,4 +1,4 @@
-from typing import Dict
+from typing import Dict, Iterator, Optional
 
 
 def format_list_of_words(first: str, *rest: str):
@@ -32,3 +32,42 @@ def format_plural_str(formatstr: str, pluralizable: Dict[str, str], use_plural: 
         for singular, plural in pluralizable.items()
     }
     return formatstr.format(**argdict)
+
+
+class CLIStubResponse:
+    """
+    A stub response class to make arbitrary data accessible in a way similar to a
+    GlobusHTTPResponse object.
+    """
+
+    def __init__(self, data):
+        self.data = data
+
+    def __getitem__(self, key):
+        return self.data[key]
+
+
+# wrap to add a `has_next()` method and `limit` param to a naive iterator
+class PagingWrapper:
+    def __init__(self, iterator: Iterator, limit: Optional[int] = None):
+        self.iterator = iterator
+        self.next = None
+        self.limit = limit
+        self._step()
+
+    def _step(self):
+        try:
+            self.next = next(self.iterator)
+        except StopIteration:
+            self.next = None
+
+    def has_next(self):
+        return self.next is not None
+
+    def __iter__(self):
+        yielded = 0
+        while self.has_next() and (self.limit is None or yielded < self.limit):
+            cur = self.next
+            self._step()
+            yield cur
+            yielded += 1
